new_branch2-03-12-25 19:19

+++ orders/models.py
@@ -28,25 +28,33 @@ class Order(models.Model):
     status = models.SmallIntegerField(default=CREATED, choices=STATUSES)
     initiator = models.ForeignKey(User, on_delete=models.CASCADE, null=True, blank=True)
 
-
+    class Meta:
+        ordering = ['-created']
     def __str__(self):
         return f'Заказ №{self.id} от {self.first_name} {self.last_name}' if self.id else 'Новый заказ'
 
-
     def update_after_payment(self):
         baskets = Basket.objects.filter(user=self.initiator)
+
         if not baskets.exists():
             logger.warning("Корзины для пользователя не найдены")
             return
+
+        # Дополнительная проверка: есть ли товары в корзинах
+        if not any(basket.items.exists() for basket in baskets):
+            logger.warning("Корзины найдены, но товаров в них нет")
+            return
+
         self.status = self.PAID
 
         self.basket_history = {
-            'purchased_items': [basket.de_json() for basket in baskets],
+            'purchased_items': [basket.de_json() for basket in baskets if basket.items.exists()],
             'total_sum': float(baskets.total_sum()),
         }
 
         baskets.delete()
         self.save()
         logger.info("Оплата обработана, корзины удалены")
+
